Helpers adjusted, ApiUrls renamed back to codesAPI
