+++ src/App/common/MovieTile/index.js
@@ -12,8 +12,8 @@ import {
   DescriptionBox,
   PosterIMG,
 } from "./styled";
-import { genres } from "../genreIds";
-import { IMG_URL } from "../Helpers/ApiUrls";
+import { genres } from "../Helpers/genreIds";
+import { IMG_URL } from "../Helpers/codesAPI";
 import whenNoPoster from "../../images/whenNoPoster.png";
 
 const MovieTile = ({ posterPath, title, year, tag, rating, votes }) => {
